Return a list of shelters.

+++ database/migrations/2018_03_10_001222_create_organization_table.php
@@ -22,8 +22,8 @@ class CreateOrganizationTable extends Migration
             $t->string('phone_number');
             $t->string('website');
             $t->string('type');
-            $t->float('lat');
-            $t->float('long');
+            $t->double('lat');
+            $t->double('long');
 
             $t->timestamps();
         });
